+++ modules/apps/portal-vulcan/portal-vulcan-test/src/testIntegration/java/com/liferay/portal/vulcan/internal/jaxrs/context/provider/test/PaginationContextProviderTest.java
@@ -12,7 +12,7 @@ import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
 import com.liferay.portal.vulcan.internal.jaxrs.context.provider.test.util.MockFeature;
 import com.liferay.portal.vulcan.internal.jaxrs.context.provider.test.util.MockMessage;
 import com.liferay.portal.vulcan.internal.jaxrs.context.provider.test.util.MockResource;
-import com.liferay.portal.vulcan.internal.util.PaginationConfigurationTestUtil;
+import com.liferay.portal.vulcan.internal.test.util.PaginationConfigurationTestUtil;
 import com.liferay.portal.vulcan.pagination.Pagination;
 import com.liferay.portal.vulcan.resource.EntityModelResource;
